+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/MsSql/MsSqlAsyncLegacyAspPipelineTests.cs
@@ -11,7 +11,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.UnboundedIntegrationTests
+namespace NewRelic.Agent.UnboundedIntegrationTests.MsSql
 {
     [NetFrameworkTest]
     public class MsSqlAsyncLegacyAspPipelineTests : IClassFixture<RemoteServiceFixtures.MsSqlBasicMvcFixture>
